Support: split out the `WindowsHandle` extensions
This centralizes the definitions for the handling of Windows Handles to
make them easier to find.
