+++ Sources/UI/Font.swift
@@ -7,17 +7,6 @@
 
 import WinSDK
 
-extension HFONT__: HandleValue {
-  typealias HandleType = HFONT
-  internal static func release(_ hFont: HandleType?) {
-    if let hFont = hFont {
-      DeleteObject(hFont)
-    }
-  }
-}
-
-internal typealias FontHandle = ManagedHandle<HFONT__>
-
 // Values derived from the Apple Human Inteface Guidelines
 // https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/typography
 private typealias TypographyInfo = (weight: Font.Weight, size: Double, leading: Int)
